+++ src/main/java/magikareborn/common/blocks/SpellAltarBlock.java
