+++ src/main/java/magikareborn/common/helpers/BlockHelper.java
