+++ src/main/java/magikareborn/common/helpers/CapabilityHelper.java
