+++ src/main/java/magikareborn/common/helpers/LightHelper.java
