+++ src/main/java/magikareborn/common/helpers/ResourceHelper.java
