+++ src/main/java/magikareborn/common/helpers/ToolHelper.java
