+++ src/main/java/magikareborn/common/items/SpellItemBase.java
