+++ src/main/java/magikareborn/common/util/RegistryHandler.java
